Revert "[Jobs] Move JobSupervisor and JobLogStorageClient to separate files  (#45116)"
This reverts commit 4ac54f10d1c8c88015e1a682ea8a068ba9938045.
